% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_intervals.R
\name{class_intervals}
\alias{class_intervals}
\title{Form class intervals}
\usage{
class_intervals(data_matrix, locations, n_class_intervals)
}
\arguments{
\item{data_matrix}{data matrix of proportions, ordered on location}

\item{locations}{Numeric vector. Locations returned from \code{estimate}}

\item{n_class_intervals}{Integer.  Number of class intervals}
}
\value{
A list of numeric matrices, one matrix for each performance, showing
  the performances it was involved with, and for each of these, the
  proportion, location and class interval used for Person Characteristic
  Curve.
}
\description{
Form class intervals
}
