% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anchor.R
\name{estimate_anch}
\alias{estimate_anch}
\title{Newton-Raphson estimation Bradley-Terry-Luce}
\usage{
estimate_anch(
  x,
  extremes,
  betas,
  adjust_extreme = 0.25,
  convergence_criteria = c(0.001, 0.001),
  loop_size = c(30, 100)
)
}
\arguments{
\item{x}{full data matrix of comparisons}

\item{extremes}{Character vector of extreme performance names (returned as an
attribute from \code{remove_xtrms}.}

\item{betas}{Numeric vector of betas estimated from \code{estimate}}

\item{loop_size}{Numeric vector length 2.  Inner maximum loops and outer
maximum loops.}

\item{convergence}{Numeric vector length 2.  Inner convergence criteria and
outer convergence criteria.}
}
\description{
Newton-Raphson estimation Bradley-Terry-Luce
}
