% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{gtpa18}
\alias{gtpa18}
\title{Graduate Teacher Performance Assessment 2018 Pairwise Comparison Project}
\format{A data frame of pairwise comparisons}
\usage{
data(gtpa18)
}
\description{
Graduate Teacher Performance Assessment 2018 Pairwise Comparison Project
}
\keyword{datasets}
