% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_matrix.R
\name{has_extremes}
\alias{has_extremes}
\title{detect extremes}
\usage{
has_extremes(x)
}
\arguments{
\item{x}{data matrix of pairwise comparison proportions or counts}
}
\value{
TRUE if any columns or rows of \code{x} contain only 0 or NA, FALSE
  otherwise.
}
\description{
detect extremes
}
