% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/judge_lmf.R
\name{judge_lmf}
\alias{judge_lmf}
\title{Fit the Logistic Measurement Function}
\usage{
judge_lmf(x, convergence_criteria = c(0.01, 0.01, 0.01, 0.01, 0.001),
  max_iterations = c(outer_loop = 20, beta_loop = 20, alpha_loop = 20,
  inner_loop = 20))
}
\arguments{
\item{x}{3D data matrix (performances \* performances \* judges)}

\item{convergence_criteria}{Stopping criteria}

\item{max_iterations}{Maximum iterations}
}
\description{
Quantify performance locations and discrimination for judges.
}
