% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outfit.R
\name{outfit}
\alias{outfit}
\title{Item Outfit}
\usage{
outfit(x, beta, type = "performance")
}
\arguments{
\item{x}{4D data array of binary pairwise comparison outcomes.}

\item{beta}{Numeric vector of location estimates for each performance}

\item{type}{Outfit statistic type: "performance", "judge", or "criterion".}
}
\value{
Data frame of names and outfit statistics.
}
\description{
Return outfit test statistics for each performance.
}
\examples{
x.dm <- pairs_format(sim_judgements_small)
x.array <- pairs_array(sim_judgements_small)
performance_estimates <- estimate_betas(x.dm)
outfit(x.array, performance_estimates$b)

}
