% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{pairs_format}
\alias{pairs_format}
\title{create a data matrix from a set of pairwise comparisons}
\usage{
pairs_format(x)
}
\arguments{
\item{x}{A data frame containing the raw results from pairwise comparisons}
}
\value{
Fulll data matrix of responses.
}
\description{
data matrix returned for estimation.
}
