% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_array.R
\name{pairs_format2}
\alias{pairs_format2}
\title{Create data matrix for estimation}
\usage{
pairs_format2(x, judges = "all", criteria = "all",
  sum_across = c("judges", "criteria"))
}
\arguments{
\item{x}{data.frame containing the variables \code{Judge, Criteria, Item,
Item.1, Selected}.}

\item{judges}{Numeric or character vector specifying judges to include.}

\item{criteria}{Numeric or character vector specifying criteria to include.}

\item{sum_across}{Character vector. \code{"judges"}, \code{"criteria"} or
\code{c("judges", "criteria")}}
}
\value{
A data matrix (2D) if \code{sum_across = "judges"} or
  \code{sum_across = "criteria"}, or if \code{sum_across = c("judges",
  "criteria")} then a 3D array is returned.
}
\description{
Sum across criteria to estimate judge discrimination, and sum across judges
to estimate criteria discrimination.
}
