% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCCs.R
\name{plot_JCCs}
\alias{plot_JCCs}
\title{Judge Characteristic Curves}
\usage{
plot_JCCs(comparisons, estimates, class_intervals, sub_title = NULL)
}
\arguments{
\item{comparisons}{Data frame of comparisons with variables \code{Judge,
Item, Item.1, Selected} and optionally \code{Criteria}.}

\item{estimates}{Data frame which inclues performance names and their
estimated location in logits.}

\item{class_intervals}{Integer specifying the number of class intervals to
form.}

\item{sub_title}{Character vector, the same length as the number of plot to
return, to use as a sub-title in the plots.}
}
\description{
Plot left/right difference in logits against proportion of left selected for
each class interval.  Overlay probability curve.
}
\references{
Steve Humphry. Personal correspondance.
}
