% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCCs.R
\name{plot_PCCs}
\alias{plot_PCCs}
\title{Plot Person Characteristic Curves}
\usage{
plot_PCCs(data_matrix, betas, class_intervals = 4,
  performances = "all")
}
\arguments{
\item{data_matrix}{Square data frame containing pairwise comparison data
converted to proportions.  Null comparisons must be represented as NA.}

\item{betas}{Named vector of locations of performance ability, in logits.}

\item{class_intervals}{The number of class intervals.}

\item{performances}{Which performances to plot.  Character "all" for all performances, or a
numeric (integer) vector of performance numbers to plot.}
}
\description{
Plot Person Characteristic Curves
}
\examples{
plot_PCCs(gtpa2018_proportions_dm, gtpa2018_locations,
          class_intervals = 3,
          performances = 21:22)

}
