% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data_matrix.R
\name{remove_xtrms}
\alias{remove_xtrms}
\title{Remove extremes}
\usage{
remove_xtrms(data_matrix)
}
\arguments{
\item{data_matrix}{full data matrix, possibly containing extremes.}
}
\description{
Remove nested extremes.  High and low extremes returned as
attributes of the reduced data matrix.
}
