% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_array.R
\name{return_criteria}
\alias{return_criteria}
\title{List all criteria}
\usage{
return_criteria(x)
}
\arguments{
\item{x}{data.frame containing the variable \code{Criteria}}
}
\value{
Character vector of all criteria names in the data set.
}
\description{
List all criteria
}
