% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_array.R
\name{return_judges}
\alias{return_judges}
\title{List all judges}
\usage{
return_judges(x)
}
\arguments{
\item{x}{data.frame containing the variable \code{Judge}}
}
\value{
Character vector of all judge names in the data set.
}
\description{
List all judges
}
