% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_examples.R
\docType{data}
\name{sim_judgements}
\alias{sim_judgements}
\title{Simulated pairwise judgements}
\format{
A data frame of pairwise comparisons
}
\usage{
data(gtpa18)
}
\description{
A simulated data set with 20 performances, 4 judges, and 2 criteria
}
\keyword{datasets}
