% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startPW.R
\name{startPW}
\alias{startPW}
\title{Perform Pairwise Comparisons Analysis and Produce PCCs}
\usage{
startPW(hold_plot = FALSE)
}
\description{
Interactive use for pairwise formating, estimation, and fit statistics.
}
