% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate2}
\alias{estimate2}
\title{Newton-Raphson estimation - faster?
Bradley-Terry-Luce}
\usage{
estimate2(x, convergence_criteria = c(0.001, 0.001), loop_size = c(30, 100))
}
\arguments{
\item{x}{data matrix of proportions}

\item{loop_size}{Numeric vector length 2.  Inner maximum loops and outer
maximum loops.}

\item{convergence}{Numeric vector length 2.  Inner convergence criteria and
outer convergence criteria.}
}
\description{
Newton-Raphson estimation - faster?
Bradley-Terry-Luce
}
