% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_betas}
\alias{estimate_betas}
\title{Estimate BTL Parameters}
\usage{
estimate_betas(x, adjust_extreme = 0.25)
}
\arguments{
\item{x}{data matrix of pairwise comparisons}

\item{adjust_extreme}{Adjustment for extremes}
}
\description{
Estimate ability. \code{estimate_betas} estimates non-extreme performances
first, then estimates extreme performances one level at a time. If the data
contain many levels of nested extremes, then \code{estimate_betas} will
likely take a long time for estimates to converge, and may fail.
}
