% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate_betas}
\alias{estimate_betas}
\title{Estimate BTL Parameters}
\usage{
estimate_betas(x, adjust_extreme = 0.25)
}
\arguments{
\item{x}{data matrix of pairwise comparisons}

\item{adjust_extreme}{Adjustment for extremes}
}
\description{
Estimate ability.
}
